+++ tests/risks/test_data_transformations.py
@@ -9,9 +9,6 @@ from vivarium_public_health.risks.data_transformations import (
     get_relative_risk_data,
 )
 from vivarium_public_health.risks.effect import RiskEffect
-from vivarium_public_health.risks.implementations.low_birth_weight_and_short_gestation import (
-    LBWSGDistribution,
-)
 from vivarium_public_health.utilities import EntityString, TargetString
 
 
@@ -108,31 +105,3 @@ def test__subset_relative_risk_to_empty_dataframe(base_config, base_plugins):
     with pytest.raises(ValueError, match=error_msg):
         get_relative_risk_data(sim._builder, risk_effect.risk, target)
 
-
-@pytest.mark.parametrize(
-    "description, expected_age_values, expected_weight_values",
-    [
-        (
-            "Neonatal preterm and LBWSG (estimation years) - [0, 24) wks, [0, 500) g",
-            (0.0, 24.0),
-            (0.0, 500.0),
-        ),
-        (
-            "Neonatal preterm and LBWSG (estimation years) - [40, 42+] wks, [2000, 2500) g",
-            (40.0, 42.0),
-            (2000.0, 2500.0),
-        ),
-        (
-            "Neonatal preterm and LBWSG (estimation years) - [34, 36) wks, [4000, 9999] g",
-            (34.0, 36.0),
-            (4000.0, 9999.0),
-        ),
-    ],
-)
-def test_parsing_lbwsg_descriptions(description, expected_weight_values, expected_age_values):
-    weight_interval = LBWSGDistribution._parse_description("birth_weight", description)
-    age_interval = LBWSGDistribution._parse_description("gestational_age", description)
-    assert weight_interval.left == expected_weight_values[0]
-    assert weight_interval.right == expected_weight_values[1]
-    assert age_interval.left == expected_age_values[0]
-    assert age_interval.right == expected_age_values[1]
